Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,25 +1,25 @@
 #include <stdio.h>
 int main() {
-    float a,b,s,p,y,x;
+    int a,b,s,p,y,x;
     char c;
     scanf("%d %d %c",&a,&b,&c);
     switch(c){
         case '+':
             s=a+b;
-            printf("%f",s);
+            printf("%d",s);
             break;
         case '-':
             p=a-b;
-            printf("%f",p);
+            printf("%d",p);
             break;
         case '*':
             x=a*b;
-            printf("%f",x);
+            printf("%d",x);
             break;
         case '/':
             y=a/b;
             if(b!=0){
-                printf("%f",y);}
+                printf("%d",y);}
             else{
                 printf("error");}
             break;
